flags -cspdi working proprely

+++ head/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: grebrune <grebrune@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/20 19:25:27 by grebrune          #+#    #+#             */
-/*   Updated: 2023/11/28 12:43:00 by grebrune         ###   ########.fr       */
+/*   Updated: 2023/11/28 15:58:15 by grebrune         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 #include <stdio.h>
 
 int	ft_printf(const char *str, ...);
-int	ft_find_type(const char c, va_list lst);
-int	ft_putnbr_b(ssize_t i, char *base, ssize_t *parse);
-int	ft_putnbr_addr(size_t i, char *base, ssize_t *parse);
+ssize_t	ft_find_type(const char c, va_list lst);
+ssize_t	ft_putnbr_b(ssize_t i, char *base, ssize_t *parse);
+ssize_t	ft_putnbr_addr(size_t i, char *base, ssize_t *parse);
 size_t	ft_putstr(char *str);
 int	ft_putchar(char c);
